requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.20
langchain-google-genai==1.0.3
langchain-community==0.0.38
google-generativeai==0.5.4
python-dotenv==1.0.0
pydantic==2.7.4
httpx==0.25.2
beautifulsoup4==4.12.2
duckduckgo-search==5.3.0